+++ next.config.js
@@ -1,12 +1,5 @@
-const withMdxEnhanced = require("next-mdx-enhanced");
-const rehypePrism = require("@mapbox/rehype-prism");
-
-module.exports = withMdxEnhanced({
-  layoutPath: "src/layouts",
-  defaultLayout: true,
-  rehypePlugins: [rehypePrism],
-})({
-  pageExtensions: ["mdx", "tsx"],
+module.exports = {
+  pageExtensions: ["tsx"],
   webpack: (config, { buildId, dev, isServer, defaultLoaders, webpack }) => {
     config.module.rules.push(
       ...[
@@ -23,4 +16,4 @@ module.exports = withMdxEnhanced({
     );
     return config;
   },
-});
+};
