+++ package.json
@@ -19,7 +19,7 @@
     "gray-matter": "^4.0.2",
     "js-yaml": "^3.14.0",
     "next": "9.5.1",
-    "next-mdx-enhanced": "^3.1.0",
+    "next-mdx-remote": "^2.1.3",
     "normalize.css": "^8.0.1",
     "react": "16.13.1",
     "react-dom": "16.13.1",
